Theres a LOT of changes in this one.
Rehauled the SpriteAtlas object, it's now seperated entirely from the json convention. There's a function for that now that can be expanded upon.

Added AnimatedSprite objects. These use a SpriteAtlas to create animations.

Added AnimatedObject. This is a child of GameObject that uses AnimatedSprites. This will likely be expanded upon in the future to account for actors that have multiple animation sequences.

Mesh now has seperate draw calls to account for UV offsets and UV scales. This is still a work in progress and will be refined later.

Added Level object. We'll eventually load a file here to create levels. For now it can only create a default map of floor.

Added Tile. Basically just a TextureObject with a predefined sprite to use.

My next goals here will be working on attributes for environmental objects for the game to observe (might do components again).