Added A* pathfinding to AI_Patterns
Added Skeleton enemy that uses the new A* Pattern

Added a TextMesh class that inherits from Mesh. It takes in a string and automatically creates a mesh to fit that string using a default Font. It can be changed as needed.

Added a TextObject base class to use TextMeshes.

Will be looking to overhaul the draw system to save on the numerous amounts of draw calls that happens.