Rehauled Shaders to now be managed globally Added Cameras that are also managed globally
For Shaders and Cameras there are some macro calls to easily access them from wherever.

Added new Canvas class. This class is a variation of the mesh that intakes verts in triangle strip format, converts them into triangles, and stores them internally. When ready, the VBO can be loaded and a singular giant mesh is drawn.

Levels are now using a Canvas to draw the tilemap. Massive CPU improvements.