final student build
Added a HUD_Score subobject that's hardlocked to the bottom left of the screen. It collects score gains and displays it.

Enemies now have score values that get added to the game state score when the enemy dies.

Levels now load on file (very little variance)
The game will now use savefiles if there is one available.

Particle shader and ParticleEmitter/Renderer classes added. they ALMOST work. There's something going on with the uniforms in the Renderers draw call that isn't working right. Couldn't figure it out with online help.